I have implemented several performance improvements to the attention mechanism by replacing the use of a dense T x T attention mask with lightweight 1D span metadata, based on an AI-generated critique.
The key changes are:

1.  **Dense Mask Removal:** The creation of dense T x T attention masks in `data_builder.py` has been removed. The dataloaders now provide `None` instead of a mask.

2.  **1D Span Metadata:** The model's forward pass in `model.py` now dynamically computes 1D `span_id`, `span_begin`, and `span_end` tensors based on special tokens in the input. These tensors are now created with `torch.int32` dtype to ensure type consistency within the Triton kernel.

3.  **Kernel Optimization:** The Triton attention kernel (`_flash_attn_fwd`) in `original_kernel.py` has been significantly optimized:
    -   It now uses the 1D span metadata to dynamically determine the key/value tiles to iterate over for each query tile. This keeps non-span tokens on a strictly causal path while allowing span tokens to attend to their specific island, reducing unnecessary computation.
    -   The internal masking logic has been replaced with a more efficient and correct implementation that properly enforces the desired attention semantics (causal for non-spans, bidirectional within a span) and disallows cross-span and span-to-non-span attention.

4.  **Interface Updates:** The function signatures for the attention kernels and their wrappers in `original_kernel.py` have been updated for both the forward and backward passes to use the new span metadata, ensuring consistency.

**Fixup:** I have also included a fix for the training and evaluation loops in `train_loop.py` and the data inspection script `entry.py`. The loops are updated to correctly handle the `None` value for the attention mask and to use the correct variable names when unpacking data batches.

+++ model.py
@@ -200,11 +200,11 @@ class GPTModel(nn.Module):
 
         if has_spans:
             in_span = (torch.cumsum(is_span_start.int(), dim=1) - torch.cumsum(is_span_end.int(), dim=1)) > 0
-            span_id = torch.cumsum(is_span_start.int(), dim=1)
+            span_id = torch.cumsum(is_span_start.int(), dim=1).to(torch.int32)
             span_id[~in_span] = 0  # 0 for non-span tokens as per critique
 
-            span_begin = torch.full_like(x, -1) # Using -1 as sentinel
-            span_end = torch.full_like(x, -1)   # Using -1 as sentinel
+            span_begin = torch.full_like(x, -1, dtype=torch.int32) # Using -1 as sentinel
+            span_end = torch.full_like(x, -1, dtype=torch.int32)   # Using -1 as sentinel
 
             for i in range(batch_size):
                 starts = torch.nonzero(is_span_start[i], as_tuple=True)[0]
@@ -221,9 +221,9 @@ class GPTModel(nn.Module):
                         span_end[i, start_idx:end_idx + 1] = end_idx
         else:
             # Create dummy tensors when no spans are present
-            span_id = torch.zeros_like(x)
-            span_begin = torch.full_like(x, -1)
-            span_end = torch.full_like(x, -1)
+            span_id = torch.zeros_like(x, dtype=torch.int32)
+            span_begin = torch.full_like(x, -1, dtype=torch.int32)
+            span_end = torch.full_like(x, -1, dtype=torch.int32)
 
         # Apply transformer blocks
         for block in self.blocks:
